api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "FastAPI masterclass"
authors = ["Miguel Angel <miguelangeldev@icloud.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
ruff = "^0.4.8"
pydantic = "^2.7.3"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.3.1"
alembic = "^1.13.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
output-format = "grouped"
extend-exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",   # pyflakes
    "UP",  # pyupgrade,
    "I",   # isort
    "D4",  # flake8-docstrings
    "PT",  # flake8-pytest
    "PL",  # flake8-pytest-style
    "Q",  # flake8-quotes
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = ["D401", "D417"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "double"
docstring-quotes = "double"

[tool.ruff.format]
# Use single quotes rather than double quotes.
quote-style = "single"
